create coinjoin-client-options.cpp/.h
