+++ src/rpc/coinjoin.cpp
@@ -5,6 +5,7 @@
 #include <validation.h>
 #ifdef ENABLE_WALLET
 #include <coinjoin/coinjoin-client.h>
+#include <coinjoin/coinjoin-client-options.h>
 #endif // ENABLE_WALLET
 #include <coinjoin/coinjoin-server.h>
 #include <rpc/server.h>
