+++ src/wallet/rpcwallet.cpp
@@ -31,6 +31,7 @@
 #include <init.h>  // For StartShutdown
 
 #include <coinjoin/coinjoin-client.h>
+#include <coinjoin/coinjoin-client-options.h>
 #include <llmq/quorums_chainlocks.h>
 #include <llmq/quorums_instantsend.h>
