+++ src/wallet/wallet.cpp
@@ -31,6 +31,7 @@
 #include <wallet/walletutil.h>
 
 #include <coinjoin/coinjoin-client.h>
+#include <coinjoin/coinjoin-client-options.h>
 #include <governance/governance.h>
 #include <keepass.h>
 #include <spork.h>
